BUG: Fix the "no data to glyph" error in glyph pipeline.
During processing the 'vtkMRMLScene::NodeAddedEvent' event, vtkMRMLModelDisplayableManager::GetActiveScalarName() is called. This function triggers an Update() on the pipeline(polydata). But at this point, not all the required field data for algorithm vtkDiffusionTensorGlyph have been filled. So the glyph pipeline complains with the error messages.
The Update() on the polydata in vtkMRMLModelDisplayableManager::GetActiveScalarName() is redundant (introduced in the fix of #2392). The Update() on the polydata in vtkMRMLModelNode::SetActivePointScalars is enough to fix #2392.

Fixes #3349